More improvements regarding blood castle event
* Abstracted similar IAttackable implementations of Monster and Destructible into a new common base class AttackableNpcBase.
* Defined the archangel weapon drop as DropItemGroup in the definition of the Crystal Statue
* Fixed that monsters were not walking into previously blocked terrain coordinates.
* Added the possibility to override the maximum health at the spawn area. We need this to apply different health values for the castle door and statues, depending on the blood castle level.
* Fixed error in the admin panel when loading the game configuration - the spawn area must not have a set GameMap, because the SpawnArea is marked with the MemberOfAggregateAttribute here, and the spawn areas of a map are marked as well. However, there can only be one "owner", otherwise it will be load twice into two objects with the same id.
