+++ src/Persistence/Initialization/VersionSeasonSix/Maps/BloodCastle5.cs
@@ -28,7 +28,7 @@ internal class BloodCastle5 : BloodCastleBase
     protected override byte MapNumber => 15;
 
     /// <inheritdoc/>
-    protected override string MapName => "Blood Castle 5";
+    protected override int CastleLevel => 5;
 
     /// <inheritdoc/>
     protected override IEnumerable<MonsterSpawnArea> CreateMonsterSpawns()
