+++ src/Persistence/Initialization/VersionSeasonSix/NpcInitialization.cs
@@ -7,6 +7,7 @@ namespace MUnique.OpenMU.Persistence.Initialization.VersionSeasonSix;
 using MUnique.OpenMU.AttributeSystem;
 using MUnique.OpenMU.DataModel.Configuration;
 using MUnique.OpenMU.GameLogic.Attributes;
+using MUnique.OpenMU.GameLogic.MiniGames;
 using MUnique.OpenMU.Persistence.Initialization.Skills;
 
 /// <summary>
@@ -750,12 +751,20 @@ internal partial class NpcInitialization : Version095d.NpcInitialization
             def.Number = 132;
             def.Designation = "Statue of Saint";
             def.ObjectKind = NpcObjectKind.Destructible;
+            def.NumberOfMaximumItemDrops = 1;
             var attributes = new Dictionary<AttributeDefinition, float>
             {
                 { Stats.MaximumHealth, 5000000 },
             };
             def.AddAttributes(attributes, this.Context, this.GameConfiguration);
             this.GameConfiguration.Monsters.Add(def);
+            var questItemDrop = this.Context.CreateNew<DropItemGroup>();
+            questItemDrop.Chance = 1;
+            questItemDrop.Description = "Archangel Weapon (Blood Castle)";
+            questItemDrop.Monster = def;
+            questItemDrop.PossibleItems.Add(this.GameConfiguration.Items.First(item => item.IsArchangelQuestItem()));
+            def.DropItemGroups.Add(questItemDrop);
+            this.GameConfiguration.DropItemGroups.Add(questItemDrop);
         }
 
         {
